flask/requirements.txt
flask
pymongo==4.2.0
bcrypt==3.2.0
PyJWT==2.3.0
python-dotenv==0.19.2